requirements.txt
playwright==1.54.0
pytest==8.3.3
pytest-asyncio==0.23.7
python-dotenv==1.0.1
loguru==0.7.2
beautifulsoup4==4.12.3
typing-extensions==4.14.1
